rossum_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rossum-agent"
version = "0.1.0"
description = "AI agent for Rossum document processing with tools for data manipulation and visualization."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["agent", "rossum", "document-processing", "ai"]
dependencies = [
    "matplotlib>=3.10",
    "plotly>=6.0",
    "pyyaml>=6.0",
    "requests>=2.32",
    "rossum-api @ git+https://github.com/rossumai/rossum-sdk#egg=rossum-api",
    "seaborn>=0.13.2",
    "smolagents[litellm]>=1.0.0",
]

[project.optional-dependencies]
docs = [
    "myst-parser>=2.0.0",
    "sphinx>=7.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-rtd-theme>=2.0.0",
]
streamlit = [
    "streamlit>=1.28.0",
]
tests = [
    "coverage>=7.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
all = [
    "coverage>=7.0.0",
    "myst-parser>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "sphinx>=7.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-rtd-theme>=2.0.0",
    "streamlit>=1.28.0",
]
